PIO: Update pin config driver for serial console
Rather than unregistering all platform devices
upon initializing, wait and only un/register
devices as needed. This allows the serial console
to be enabled by default. The HAL is responsible
for managing the console service and UART1, as
failing to stop the service before unregistering
the device causes a crash.

To stop the service:
setprop ctl.stop console

To start the service:
setprop ctl.start console

Bug: http://b/31709012
Test: serial console works and driver tests pass
Change-Id: I5eb893537726607fb5ff16dc6cc4dbe7a746178a