PIO: Uncouple pin groups from pin functions
Fixes a big with SPI0 CS pins not working after
being unregistered. This was due to the fact that
the CS pins are set to function GPIO rather than
SPI0. Also fixes a bug with PWM interfering with
GPIO on pin 12 by allowing default device pins to
come from different groups.

Bug: http://b/62956237
Test: SPI0 and PWM work
Change-Id: I37eef9065ee68fd5d1a5d8df6c99793407acacf3